add app.js file
